refactor: share footer component
